fix(64bit): resolve the 15 -Wshorten-64-to-32 LP64 truncations (#10)
Phase 6, commit 2. Make every -Wshorten-64-to-32 site explicit so golden
stays byte-identical (the implicit conversion already truncated exactly
this way; the cast/wider-type only documents it).

- z.c + mapgen/z.c readlin path: `nread` retyped int->ssize_t (its source
  is read()), clearing both `nread = read(...)` assignment sites in each;
  downstream indexing/compares are unaffected.
- mapgen/z.c str_save: (unsigned) cast on `strlen(s) + 1` feeding the
  mapgen my_malloc(unsigned size). (olympia's my_malloc takes size_t, so
  its str_save was never flagged — divergence from G1.)
- olympia/rnd.c md5_int: `return (int) buf[0]` — low 32 bits of the MD5
  word (was a 32-bit unsigned long on ILP32).
- olympia/code.c letter_val: `return (int)(p-let)` — index into a fixed
  short string.
- strlen()->int name/line/word lengths, provably <2^31: documented (int)
  casts in z.c/mapgen z.c fuzzy_strcmp, c2.c line_length_check, check.c
  check_loc_name_lengths, eat.c do_eat_command, report.c
  strip_leading_stupid_word.

Debug and asan-ubsan golden gates both YES (byte-identical); asan/ubsan
clean.

+++ olympia/c2.c
@@ -160,7 +160,7 @@ line_length_check(cstrings_list l)
 	int len = 0;
 
 	for (i = 0; i < cstrings_len(l); i++)
-		len = max(len, strlen(l[i]));
+		len = max(len, (int) strlen(l[i]));	/* line len fits 32 bits */
 
 	return len;
 }
